@@ -10,7 +10,6 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-import typing as ty
 
 from keystoneauth1.extras import _saml2
 from keystoneauth1 import loading
@@ -19,7 +18,7 @@ from keystoneauth1.loading import opts
 
 class Saml2Password(loading.BaseFederationLoader[_saml2.V3Saml2Password]):
     @property
-    def plugin_class(self) -> ty.Type[_saml2.V3Saml2Password]:
+    def plugin_class(self) -> type[_saml2.V3Saml2Password]:
         return _saml2.V3Saml2Password
 
     @property
@@ -51,7 +50,7 @@ class Saml2Password(loading.BaseFederationLoader[_saml2.V3Saml2Password]):
 
 class ADFSPassword(loading.BaseFederationLoader[_saml2.V3ADFSPassword]):
     @property
-    def plugin_class(self) -> ty.Type[_saml2.V3ADFSPassword]:
+    def plugin_class(self) -> type[_saml2.V3ADFSPassword]:
         return _saml2.V3ADFSPassword
 
     @property
