@@ -10,7 +10,6 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-import typing as ty
 
 from keystoneauth1.extras.oauth1 import v3
 from keystoneauth1 import loading
@@ -21,7 +20,7 @@ from keystoneauth1.loading import opts
 # include the scoping options like project-id in the option list
 class V3OAuth1(loading.BaseIdentityLoader[v3.OAuth1]):
     @property
-    def plugin_class(self) -> ty.Type[v3.OAuth1]:
+    def plugin_class(self) -> type[v3.OAuth1]:
         return v3.OAuth1
 
     @property
